Refactor to allow fields and their validation to be arbitrary. Requires the markup uses same name but otherwise should allow for additional fields of various types to be added
